move components from shame -> derived. Update index and tests
