--- src/components/fields/CanvasSize.js
@@ -1,11 +1,0 @@
-import Numeric from './Numeric';
-import {connectToContainer} from '../../lib';
-
-function modifyPlotProps(props, context, plotProps) {
-  const {fullContainer} = plotProps;
-  if (plotProps.isVisible && fullContainer && fullContainer.autosize) {
-    plotProps.isVisible = false;
-  }
-}
-
-export default connectToContainer(Numeric, {modifyPlotProps});
