+++ src/components/index.js
@@ -6,6 +6,8 @@ import {
   Dropdown,
   Flaglist,
   Info,
+  LayoutNumericFraction,
+  LayoutNumericFractionInverse,
   Radio,
   DataSelector,
   Numeric,
@@ -34,6 +36,8 @@ export {
   Flaglist,
   Fold,
   Info,
+  LayoutNumericFraction,
+  LayoutNumericFractionInverse,
   Numeric,
   Panel,
   PanelMenuWrapper,
