+++ src/index.js
@@ -20,6 +20,8 @@ import {
   Fold,
   Info,
   Layout,
+  LayoutNumericFraction,
+  LayoutNumericFractionInverse,
   Numeric,
   Panel,
   PanelMenuWrapper,
@@ -46,6 +48,8 @@ export {
   Info,
   Hub,
   Layout,
+  LayoutNumericFraction,
+  LayoutNumericFractionInverse,
   Numeric,
   Panel,
   PanelMenuWrapper,
